factored JS code into 2 separate files -- starting Student modal show function
