+++ app/javascript/components/cards/card_modal.js
@@ -1,0 +1,12 @@
+import { linkClickSelector } from "@rails/ujs";
+
+const initModalShow = (cardInnerLinks) => {
+  // Retrieve student id
+  const studentId = cardInnerLinks.dataset["studentId"];
+
+  // Build student url with it
+
+  console.log(studentId)
+}
+
+export { initModalShow };
