+++ app/javascript/packs/application.js
@@ -27,11 +27,10 @@ import "bootstrap";
 
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
-import { initCardSelect } from '../components/select_cards';
+import { initCardsFeatures } from '../components/cards_features';
 
 document.addEventListener('turbolinks:load', () => {
   // Call your functions here, e.g:
   // initSelect2();
-  initCardSelect();
-
+  initCardsFeatures();
 });
