+++ app/views/pages/home.html.erb
@@ -12,7 +12,7 @@
     <% end %>
   </div>
 
-  <div class="row" id="card-container">
+  <div class="row card-container">
     <% @students.sample(42).each do |student| %>
       <div class="col col-12 col-md-6 col-lg-4">
         <div class="card-product position-relative" data-student-id=<%= student.id %>>
@@ -20,9 +20,13 @@
             <div class="card-product-infos d-flex flex-column flex-grow-1 justify-content-between">
               <div class="d-flex justify-content-between align-items-top">
                 <h2><%= student.name %></h2>
-                <%= link_to student_path(student.id), method: :get, class: '' do %>
+
+                <div class="card-inner-links" data-student-id=<%= student.id %>>
+
                   <i class="far fa-eye ml-3 pr-3"></i>
-                <% end %>
+
+                </div>
+
               </div>
               <p class="text-muted">Batch #<%= student.batch %></p>
               <p>
